shs-backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
pandas
numpy
scikit-fuzzy  # If needed for fuzzy logic, but the code uses numpy
python-multipart